API and polyfill updates
On the index.html page two new scripts were added. One to src a promise polyfill js file and the other to src a fetch polyfill js file.

On the scripts.js file the static repository of pokemons was removed and in replace I added an API to GET the pokemon list details.

the add(function) was updated to the correct object identifiers.

the loadList function was added to fetch the list of pokemons.

the loadDetails function was added to fetch the pokemon details from the pokemon items.

lastly the return was updated to call all functions.