Modal Updates
Added a modal feature to display the name of the pokemon when you click on the original button. When the modal opens up you see the name of the pokemon, the height of the pokemon, and an image. Then when you click either the close button, outside of the modal or use the esc key, the modal closes.